Modify extend types for a cleaner usage with derive
Minor external type changes, no functional breaking changes. It would be ideal if both define and extend could use the same Config, but it currently does not seem like that will work. With the previous setup, using derive with extend required doing the same DiffProperties & Partial to match the Config signature for the DerivedFunction. With this change, it can accept the entire Result type for a cleaner external API