Simplify the derive type signature
Remove the need to define the explicity type union for the fields. This leaves room for error if a specific field is not added as a dependent field, it won't be included in the resulting object. However, in favor of a slightly cleaner/less verbose API it was necessary to remove to keep it usable